+++ src/officialeye/interpretation/methods/__init__.py
@@ -1,0 +1,3 @@
+"""
+Package containing all intepretation methods built into OfficialEye.
+"""